Docs: enable badges; update wording
Enable badges

Gemnasium has been closed since it was acquired by GitLab, thus it is
removed. The build and coverage links don't work, but that might be
caused by an configuration issue, so they are left as disabled for now.

https://docs.gitlab.com/ee/user/project/import/gemnasium.html

Update wording regarding phantomjs support

While old version does not support Function.prototype.bind, the latest
(2.1.1) does. Phantomjs is also no longer in development, so it's safe
to assume it will stay the same. Even if it becomes active, it most
likely won't go back to an older, unsupporting, version of QT/WebKit.

http://phantomjs.org/release-2.1.html
https://github.com/ariya/phantomjs/issues/15344